pyproject.toml
[tool.poetry]
name = "meeting_sample"
version = "0.1.0"
description = ""
authors = ["刘凤一 <liufengyi@jjl.cn>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.6"
drf-yasg = "^1.20.0"
djangorestframework-jwt = "^1.11.0"
daphne = "^3.0.2"
django-safedelete = "^1.0.0"
python-dotenv = "^0.19.0"
dj-database-url = "^0.5.0"
mysqlclient = "^2.0.3"
pycryptodome = "^3.10.1"
redis-py-cluster = "^2.1.3"
djangorestframework-camel-case = "^1.2.0"
sentry-sdk = "^1.3.1"

[tool.poetry.dev-dependencies]
rust = "^0.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
